* merged all *_impl headers into their declaring headers * including mallocMC.hpp instead of individual headers in tests and examples
