+++ examples/mallocMC_example01.cpp
@@ -28,12 +28,7 @@
 
 #include <alpaka/alpaka.hpp>
 #include <iostream>
-#include <mallocMC/AlignmentPolicies.hpp>
-#include <mallocMC/CreationPolicies.hpp>
-#include <mallocMC/DistributionPolicies.hpp>
-#include <mallocMC/OOMPolicies.hpp>
-#include <mallocMC/ReservePoolPolicies.hpp>
-#include <mallocMC/mallocMC_hostclass.hpp>
+#include <mallocMC/mallocMC.hpp>
 #include <numeric>
 
 using Dim = alpaka::dim::DimInt<1>;
