+++ src/include/mallocMC/alignmentPolicies/Shrink.hpp
@@ -31,10 +31,32 @@
 
 #pragma once
 
+#include "Shrink.hpp"
+
+#include <alpaka/core/Common.hpp>
+#include <cstdint>
+#include <iostream>
+#include <sstream>
+#include <string>
+
 namespace mallocMC
 {
     namespace AlignmentPolicies
     {
+        namespace Shrink2NS
+        {
+            template<int PSIZE>
+            struct __PointerEquivalent
+            {
+                using type = unsigned int;
+            };
+            template<>
+            struct __PointerEquivalent<8>
+            {
+                using type = unsigned long long;
+            };
+        } // namespace Shrink2NS
+
         namespace ShrinkConfig
         {
             struct DefaultShrinkConfig
@@ -56,7 +78,81 @@ namespace mallocMC
          * @tparam T_Config (optional) The alignment to use
          */
         template<typename T_Config = ShrinkConfig::DefaultShrinkConfig>
-        class Shrink;
+        class Shrink
+        {
+        public:
+            using Properties = T_Config;
+
+        private:
+            using PointerEquivalent
+                = Shrink2NS::__PointerEquivalent<sizeof(char *)>::type;
+
+/** Allow for a hierarchical validation of parameters:
+ *
+ * shipped default-parameters (in the inherited struct) have lowest precedence.
+ * They will be overridden by a given configuration struct. However, even the
+ * given configuration struct can be overridden by compile-time command line
+ * parameters (e.g. -D MALLOCMC_AP_SHRINK_DATAALIGNMENT 128)
+ *
+ * default-struct < template-struct < command-line parameter
+ */
+#ifndef MALLOCMC_AP_SHRINK_DATAALIGNMENT
+#define MALLOCMC_AP_SHRINK_DATAALIGNMENT (Properties::dataAlignment)
+#endif
+            static constexpr size_t dataAlignment
+                = MALLOCMC_AP_SHRINK_DATAALIGNMENT;
+
+            // dataAlignment must be a power of 2!
+            static_assert(
+                dataAlignment != 0
+                    && (dataAlignment & (dataAlignment - 1)) == 0,
+                "dataAlignment must also be a power of 2");
+
+        public:
+            static auto alignPool(void * memory, size_t memsize)
+                -> std::tuple<void *, size_t>
+            {
+                PointerEquivalent alignmentstatus
+                    = ((PointerEquivalent)memory) & (dataAlignment - 1);
+                if(alignmentstatus != 0)
+                {
+                    std::cout << "Heap Warning: memory to use not "
+                              << dataAlignment << " byte aligned...\n"
+                              << "Before:\n"
+                              << "dataAlignment:   " << dataAlignment << '\n'
+                              << "Alignmentstatus: " << alignmentstatus << '\n'
+                              << "size_t memsize   " << memsize << " byte"
+                              << '\n'
+                              << "void *memory     " << memory << '\n';
+
+                    memory
+                        = (void *)(((PointerEquivalent)memory) + dataAlignment - alignmentstatus);
+                    memsize -= dataAlignment + (size_t)alignmentstatus;
+
+                    std::cout << "Was shrunk automatically to: " << '\n'
+                              << "size_t memsize   " << memsize << " byte"
+                              << '\n'
+                              << "void *memory     " << memory << '\n';
+                }
+
+                return std::make_tuple(memory, memsize);
+            }
+
+            ALPAKA_FN_HOST_ACC
+            static auto applyPadding(size_t bytes) -> size_t
+            {
+                constexpr auto bitsToClear = dataAlignment - 1;
+                return (bytes + bitsToClear) & ~bitsToClear;
+            }
+
+            ALPAKA_FN_HOST
+            static auto classname() -> std::string
+            {
+                std::stringstream ss;
+                ss << "Shrink[" << dataAlignment << "]";
+                return ss.str();
+            }
+        };
 
     } // namespace AlignmentPolicies
 } // namespace mallocMC
