+++ src/include/mallocMC/oOMPolicies/BadAllocException.hpp
@@ -27,6 +27,12 @@
 
 #pragma once
 
+#include "BadAllocException.hpp"
+
+#include <alpaka/core/Common.hpp>
+#include <cassert>
+#include <string>
+
 namespace mallocMC
 {
     namespace OOMPolicies
@@ -41,7 +47,31 @@ namespace mallocMC
          * accelerators that do not support exceptions results in undefined
          * behaviour.
          */
-        struct BadAllocException;
+        struct BadAllocException
+        {
+            ALPAKA_FN_ACC
+            static auto handleOOM(void * mem) -> void *
+            {
+#if BOOST_LANG_CUDA || BOOST_COMP_HIP
+//#if __CUDA_ARCH__ < 350
+#define PM_EXCEPTIONS_NOT_SUPPORTED_HERE
+//#endif
+#endif
+
+#ifdef PM_EXCEPTIONS_NOT_SUPPORTED_HERE
+#undef PM_EXCEPTIONS_NOT_SUPPORTED_HERE
+                assert(false);
+#else
+                throw std::bad_alloc{};
+#endif
+                return mem;
+            }
+
+            static auto classname() -> std::string
+            {
+                return "BadAllocException";
+            }
+        };
 
     } // namespace OOMPolicies
 } // namespace mallocMC
