--- src/include/mallocMC/oOMPolicies/BadAllocException_impl.hpp
@@ -1,67 +1,0 @@
-/*
-  mallocMC: Memory Allocator for Many Core Architectures.
-
-  Copyright 2014 Institute of Radiation Physics,
-                 Helmholtz-Zentrum Dresden - Rossendorf
-
-  Author(s):  Carlchristian Eckert - c.eckert ( at ) hzdr.de
-
-  Permission is hereby granted, free of charge, to any person obtaining a copy
-  of this software and associated documentation files (the "Software"), to deal
-  in the Software without restriction, including without limitation the rights
-  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-  copies of the Software, and to permit persons to whom the Software is
-  furnished to do so, subject to the following conditions:
-
-  The above copyright notice and this permission notice shall be included in
-  all copies or substantial portions of the Software.
-
-  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-  AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-  THE SOFTWARE.
-*/
-
-#pragma once
-
-#include "BadAllocException.hpp"
-
-#include <alpaka/core/Common.hpp>
-#include <cassert>
-#include <string>
-
-namespace mallocMC
-{
-    namespace OOMPolicies
-    {
-        struct BadAllocException
-        {
-            ALPAKA_FN_ACC
-            static auto handleOOM(void * mem) -> void *
-            {
-#if BOOST_LANG_CUDA || BOOST_COMP_HIP
-//#if __CUDA_ARCH__ < 350
-#define PM_EXCEPTIONS_NOT_SUPPORTED_HERE
-//#endif
-#endif
-
-#ifdef PM_EXCEPTIONS_NOT_SUPPORTED_HERE
-#undef PM_EXCEPTIONS_NOT_SUPPORTED_HERE
-                assert(false);
-#else
-                throw std::bad_alloc{};
-#endif
-                return mem;
-            }
-
-            static auto classname() -> std::string
-            {
-                return "BadAllocException";
-            }
-        };
-
-    } // namespace OOMPolicies
-} // namespace mallocMC
