+++ src/include/mallocMC/reservePoolPolicies/CudaSetLimits.hpp
@@ -27,6 +27,14 @@
 
 #pragma once
 
+#ifdef ALPAKA_ACC_GPU_CUDA_ENABLED
+
+#include "CudaSetLimits.hpp"
+
+#include <cuda_runtime_api.h>
+#include <mutex>
+#include <string>
+
 namespace mallocMC
 {
     namespace ReservePoolPolicies
@@ -45,7 +53,31 @@ namespace mallocMC
          * is reset via cudaDeviceReset(). See:
          * https://docs.nvidia.com/cuda/cuda-runtime-api/group__CUDART__DEVICE.html#group__CUDART__DEVICE_1g05956f16eaa47ef3a4efee84563ccb7d
          */
-        struct CudaSetLimits;
+        // TODO alpaka
+        struct CudaSetLimits
+        {
+            template<typename AlpakaDev>
+            auto setMemPool(const AlpakaDev & dev, size_t memsize) -> void *
+            {
+                cudaDeviceSetLimit(cudaLimitMallocHeapSize, memsize);
+                return nullptr;
+            }
+
+            static void resetMemPool(void * p = nullptr)
+            {
+                cudaDeviceSetLimit(cudaLimitMallocHeapSize, 8192U);
+                cudaGetLastError(); // cudaDeviceSetLimit() usually fails if any
+                                    // kernel before used ::malloc(), so let's
+                                    // clear the error state
+            }
+
+            static auto classname() -> std::string
+            {
+                return "CudaSetLimits";
+            }
+        };
 
     } // namespace ReservePoolPolicies
 } // namespace mallocMC
+
+#endif
